PCGII() is the function to apply the proposed method to get the estimated partial correlation graph with information incorporation

PCGII() is the function to apply the proposed method to get the estimated partial correlation graph with information incorporation

PCGII(df, prior, lambda)

Arguments

df

The main expression dataset, an n by p matrix, in which each row corresponds to a sample and each column represents expression/abundance of an omics feature.

prior

The prior set, a k by 2 dataframe, in which each row corresponds to a pair of nodes (any omics features) that are connected under prior belief. Note, prior input has to be dataframe.

lambda

The regularization parameter, used in the node-wise regression. If missing, default lambda will be used which is at the order of sqrt(2*log(p)/n).

Value

A list. The list contains estimated partial correlation matrix (Est), sparse partial correlation estimation matrix with threshold (EstThresh), estimated kappa (kappa), estimated test statistics matrix of partial correlations (tscore), sample size (n) and number of nodes (p).

Examples

library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union
library(tidyverse)
#> Warning: package ‘lubridate’ was built under R version 4.3.1
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ dplyr     1.1.2     ✔ readr     2.1.4
#> ✔ forcats   1.0.0     ✔ stringr   1.5.0
#> ✔ ggplot2   3.4.2     ✔ tibble    3.2.1
#> ✔ lubridate 1.9.3     ✔ tidyr     1.3.0
#> ✔ purrr     1.0.2     
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ lubridate::%--%()      masks igraph::%--%()
#> ✖ dplyr::as_data_frame() masks tibble::as_data_frame(), igraph::as_data_frame()
#> ✖ purrr::compose()       masks igraph::compose()
#> ✖ tidyr::crossing()      masks igraph::crossing()
#> ✖ dplyr::filter()        masks stats::filter()
#> ✖ dplyr::lag()           masks stats::lag()
#> ✖ purrr::simplify()      masks igraph::simplify()
#> ℹ Use the conflicted package (<http://conflicted.r-lib.org/>) to force all conflicts to become errors
library(PCGII)
# Simulating data
set.seed(1234567)
n=50 # sample size
p=30 # number of nodes

omega=make_random_precision_mat(eta=.01, p=p)

# population covariance matrix, which is used to generate data
Sigma=solve(omega)
# simulate expression data
X = rmvnorm(n = n, sigma = Sigma)
#> Error in rmvnorm(n = n, sigma = Sigma): could not find function "rmvnorm"

lam=2*sqrt(log(p)/n) ## fixed lambda

# directed prior network
prior_set=as.data.frame(matrix(data=c(5,6, 28,24), nrow=2, ncol=2, byrow = TRUE))
colnames(prior_set)=c("row", "col")
prior_set=undirected_prior(prior_set)
PCGII_out=PCGII(df=X, prior=prior_set, lambda = lam)
#> Loading required package: glmnet
#> Loading required package: Matrix
#> Warning: package ‘Matrix’ was built under R version 4.3.1
#> 
#> Attaching package: ‘Matrix’
#> The following objects are masked from ‘package:tidyr’:
#> 
#>     expand, pack, unpack
#> Loaded glmnet 4.1-7
#> Error in eval(expr, envir, enclos): object 'X' not found
inference_out=inference(list=PCGII_out)
#> Error in eval(expr, envir, enclos): object 'PCGII_out' not found
diag(inference_out)=0
#> Error: object 'inference_out' not found
net=inference_out %>% graph_from_adjacency_matrix(mode = "undirected")
#> Error in eval(expr, envir, enclos): object 'inference_out' not found
net %>% plot(vertex.size=4, vertex.label.dist=0.5, vertex.color="red", edge.arrow.size=0.5, layout=layout_in_circle(net))
#> Error in eval(expr, envir, enclos): object 'net' not found
## Remark: mathematical standardization will be automatically done within the function.